Generator/requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
basicsr==1.4.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
diffusers==0.28.2
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.64.1
huggingface-hub==0.23.3
idna==3.4
imageio==2.34.1
importlib_metadata==7.1.0
intel-openmp==2021.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
opencv-python==4.10.0.82
packaging==24.0
peft==0.11.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
realesrgan==0.3.0
regex==2024.5.15
requests==2.28.1
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
sympy==1.12
tb-nightly==2.17.0a20240605
tbb==2021.11.0
tensorboard-data-server==0.7.2
tifffile==2024.5.22
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0+cu118
torchvision==0.15.2+cu118
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.9.0
urllib3==1.26.13
Werkzeug==3.0.3
xformers==0.0.26.post1+cu118
yapf==0.40.2
zipp==3.19.2
